requirements.txt
Django==4.0.4
Pillow==9.1.0
django-ckeditor==6.4.0